refactor: refactor cda-sovd bulk data into module
